+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 Breaking:
   - RxQuery's are now [immutable](docs/RxQuery.md)
+  - RxQuery.$ does not emit `null` when running
+
 
 ## 3.0.7 (April 10, 2017)